winkernel: add service.go to hold service info
